[material-app-rating] implemented custom rating bar functionality

+++ app-rating/src/main/java/com/stepstone/apprating/AppRatingDialog.java
@@ -64,7 +64,7 @@ public class AppRatingDialog {
 
             int numberOfStars = C.InitialValues.MAX_RATING;
 
-            float defaultRating = C.InitialValues.DEFAULT_RATING;
+            int defaultRating = C.InitialValues.DEFAULT_RATING;
 
             String positiveButtonText;
